feat: print jwt header (#37)
- refactor: access/refresh token processing in jwt_token.go
